+++ src/routes/division.js
@@ -247,12 +247,12 @@ const DivisionsPending = memo(({ doReload }) => {
 });
 
 const DivisionManagers = memo(() => {
-    const { users, memberUIDs } = useContext(AppContext);
+    const { allPerms, users, memberUIDs } = useContext(AppContext);
     const allMembers = memberUIDs.map((uid) => users[uid]);
 
     let managers = [];
     for (let i = 0; i < allMembers.length; i++) {
-        if (checkPerm(allMembers[i].roles, ["administrator", "manage_divisions"])) {
+        if (checkPerm(allMembers[i].roles, ["administrator", "manage_divisions"], allPerms)) {
             managers.push(allMembers[i]);
         }
     }