+++ src/routes/downloads.js
@@ -243,12 +243,12 @@ const DownloadableItemGrid = memo(({ downloadableItems, lastUpdate, onEdit, onDe
 });
 
 const DownloadableItemManagers = memo(() => {
-    const { users, memberUIDs } = useContext(AppContext);
+    const { allPerms, users, memberUIDs } = useContext(AppContext);
     const allMembers = memberUIDs.map((uid) => users[uid]);
 
     let managers = [];
     for (let i = 0; i < allMembers.length; i++) {
-        if (checkPerm(allMembers[i].roles, ["administrator", "manage_downloads"])) {
+        if (checkPerm(allMembers[i].roles, ["administrator", "manage_downloads"], allPerms)) {
             managers.push(allMembers[i]);
         }
     }