+++ src/routes/poll.js
@@ -416,12 +416,12 @@ const PollGrid = memo(({ polls, lastUpdate, onEdit, onDelete, onPollVoters }) =>
 });
 
 const PollManagers = memo(() => {
-    const { users, memberUIDs } = useContext(AppContext);
+    const { allPerms, users, memberUIDs } = useContext(AppContext);
     const allMembers = memberUIDs.map((uid) => users[uid]);
 
     let managers = [];
     for (let i = 0; i < allMembers.length; i++) {
-        if (checkPerm(allMembers[i].roles, ["administrator", "manage_polls"])) {
+        if (checkPerm(allMembers[i].roles, ["administrator", "manage_polls"], allPerms)) {
             managers.push(allMembers[i]);
         }
     }