+++ changelog.md
@@ -54,6 +54,7 @@
 20. Moved loading of `applicationTypes` and `divisions` to "before needed" rather than "on page load"
 21. Replaced `vars.languages` with Context API `languages`, also supported background loading for supported languages
 22. Replaced `vars.roles` with Context API `allRoles`, also removed `vars.orderedRoles` which is to be calculated in UserCard
+23. Replaced `vars.perms` with Context API `allPerms`
 
 ## v3.3.x